Move webhook related helper functions back to webhook package
In addition, in order to overcome redundant looping over labels
the FilterKubevirtLabels function will also indicate whether we
have restricted labels among the kubevirt.io prefixed ones
