+++ pkg/util/BUILD.bazel
@@ -8,10 +8,5 @@ go_library(
     ],
     importpath = "kubevirt.io/kubevirt/pkg/util",
     visibility = ["//visibility:public"],
-    deps = [
-        "//pkg/virt-operator/creation/rbac:go_default_library",
-        "//staging/src/kubevirt.io/client-go/api/v1:go_default_library",
-        "//staging/src/kubevirt.io/client-go/log:go_default_library",
-        "//staging/src/kubevirt.io/client-go/util:go_default_library",
-    ],
+    deps = ["//staging/src/kubevirt.io/client-go/api/v1:go_default_library"],
 )
