+++ pkg/virt-api/webhooks/validating-webhook/admitters/vmi-update-admitter.go
@@ -20,14 +20,18 @@
 package admitters
 
 import (
+	"fmt"
 	"reflect"
+	"strings"
 
-	"kubevirt.io/kubevirt/pkg/util"
+	"kubevirt.io/client-go/log"
+	"kubevirt.io/kubevirt/pkg/virt-operator/creation/rbac"
 
 	"k8s.io/api/admission/v1beta1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	v1 "kubevirt.io/client-go/api/v1"
+	clientutil "kubevirt.io/client-go/util"
 	webhookutils "kubevirt.io/kubevirt/pkg/util/webhooks"
 )
 
@@ -70,13 +74,13 @@ func admitVMILabelsUpdate(
 	ar *v1beta1.AdmissionReview) *v1beta1.AdmissionResponse {
 
 	// Skip admission for internal components
-	allowed := util.GetAllowedServiceAccounts()
+	allowed := GetAllowedServiceAccounts()
 	if _, ok := allowed[ar.Request.UserInfo.Username]; ok {
 		return nil
 	}
 
-	oldLabels := util.FilterKubevirtLabels(oldVMI.ObjectMeta.Labels)
-	newLabels := util.FilterKubevirtLabels(newVMI.ObjectMeta.Labels)
+	oldLabels, _ := FilterKubevirtLabels(oldVMI.ObjectMeta.Labels)
+	newLabels, _ := FilterKubevirtLabels(newVMI.ObjectMeta.Labels)
 
 	if !reflect.DeepEqual(oldLabels, newLabels) {
 		return webhookutils.ToAdmissionResponse([]metav1.StatusCause{
@@ -88,4 +92,42 @@ func admitVMILabelsUpdate(
 	}
 
 	return nil
+}
+
+func GetAllowedServiceAccounts() map[string]struct{} {
+	ns, err := clientutil.GetNamespace()
+	logger := log.DefaultLogger()
+
+	if err != nil {
+		logger.Info("Failed to get namespace. Fallback to default: 'kubevirt'")
+		ns = "kubevirt"
+	}
+
+	// system:serviceaccount:{namespace}:{kubevirt-component}
+	prefix := fmt.Sprintf("%s:%s:%s", "system", "serviceaccount", ns)
+	return map[string]struct{}{
+		fmt.Sprintf("%s:%s", prefix, rbac.ApiServiceAccountName):        {},
+		fmt.Sprintf("%s:%s", prefix, rbac.HandlerServiceAccountName):    {},
+		fmt.Sprintf("%s:%s", prefix, rbac.ControllerServiceAccountName): {},
+	}
+}
+
+func FilterKubevirtLabels(labels map[string]string) (map[string]string, bool) {
+	hasRestrictedLabels := false
+	m := make(map[string]string)
+	if len(labels) == 0 {
+		// Return the empty map to avoid edge cases
+		return m, hasRestrictedLabels
+	}
+	for label, value := range labels {
+		if strings.HasPrefix(label, "kubevirt.io") {
+			if !hasRestrictedLabels {
+				if _, ok := filteredVmiKubevirtLabels[label]; ok {
+					hasRestrictedLabels = true
+				}
+			}
+			m[label] = value
+		}
+	}
+	return m, hasRestrictedLabels
 }
